calendar_agent/requirements.txt
streamlit
fastapi
uvicorn
google-api-python-client
google-auth
python-dotenv
pytz
dateparser
python-base64
